Refactor prompt stuff into single fine instead of finetune
